Move src dir
Move src dir to a more conventional location
