+++ src/lib/ractive/components.js
